# doc/api.yaml
...
openapi: 3.0.3 
info:
  title: WASAPhoto backend REST APIs
  description: |-
    Keep in touch with your friends by sharing photos of special moments, 
    thanks to WASAPhoto! You can upload your photos directly from your PC, 
    and they will be visible to everyone following you.
  version: "0.0.1"
components:
  securitySchemes:
    bearerAuth:
      description: API authentication to authorize requests
      type: http
      scheme: bearer
  responses:
    BadRequest: 
      description: |-
        The request was not compliant with the documentation
        (eg. missing fields, etc).
    InternalServerError: 
      description: |-
        The server encountered an internal error.
        Further info in server logs.
    NotFound:
      description: Page not found.
...
    UnauthorizedError:
      description: Access token is missing or invalid.
  parameters:
    username:
      schema: {$ref: "#/components/schemas/username"}
      name: username
      in: path 
      required: true
...
    userBan:
      schema: {$ref: "#/components/schemas/username"}
      name: userBan
      in: path 
      required: true
...
    photo-id:
      schema:
        description: ID of a photo
        type: integer
        example: 1
      name: photo-id
      in: path
      required: true
...
  schemas:
...
    username:
      title: Username
      description: The username
      type: string
      pattern: |-
        ^(?=.{3,16}$)[a-zA-Z0-9._]+$
      minLength: 3 
      maxLength: 16
      example: Maria 
    access-token:
      title: Authorization-token
      description: The user identifier for the authorization.
      type: object
      properties:
        identifier:
          type: string
          description: The user identifier
          example: "abcdef012345"
          pattern: ".*"
          minLength: 3
          maxLength: 16
    id:
      description: Unique identifier.
      type: integer
      example: 1
      readOnly: true
...
    photoFile:
      type: string
      format: binary
      description: The photo file
      pattern: ".*"
      minLength: 0
      maxLength: 1000000
    photoUrl:
      type: string
      pattern: "http://.*"
      description: The url of the photo
      title: URL
      minLength: 10
      maxLength: 500
      example: http://localhost:8080/imageurl18087
...
    commentArray:
      description: An array of comments
      type: array
      items: {$ref: "#/components/schemas/comment"}
      minItems: 0
      maxItems: 10
...
    comment:
      title: Comment
      description: An object with username, text and timestamp of publication of the comment.
      type: object
      properties:
        user: {$ref: "#/components/schemas/username"}
        text: {$ref: "#/components/schemas/commentText"}
        timestamp: {$ref: "#/components/schemas/timestamp"}
        image: {$ref: "#/components/schemas/photoUrl"}
        id: {$ref: "#/components/schemas/id"}
    timestamp:
      title: Timestamp
      description: The timestamp 
      type: string
      minLength: 19
      maxLength: 19
      format: date-time
      pattern: |-
        [0-9]{1,4}-[0-9]{1,2}-[0-9]{1,2}T[0-9]{1,2}:[0-9]{1,2}:[0-9]{1,2}
      example: "2022-05-23T08:00:00"
      readOnly: true
    commentText:
      type: string
      description: The text of the comment
      pattern: ".*"
      minLength: 1
      maxLength: 250
      example: Nice photo!
...
paths:
  /session/:
    post:
      tags: ["login"]
      summary: Logs in the user 
      description: |-
          If the user does not exist, it will be created,
          and an identifier is returned.
          If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details.
        content:
          application/json: 
            schema: 
              description: A username object
              type: object 
              properties:
                username: {$ref: "#/components/schemas/username"}
        required: true 
      responses:
        '201':
          description: User log-in action successful.
          content:
              application/json: 
                schema: {$ref: "#/components/schemas/access-token"}
        "400": {$ref: "#/components/responses/BadRequest"}
        '404': {$ref: '#/components/responses/NotFound'}
        "500": {$ref: "#/components/responses/InternalServerError"}
...
  /users/{username}/media/{photo-id}/:
    parameters:
      - $ref: "#/components/parameters/photo-id"
      - $ref: "#/components/parameters/username"
    
    get:
      security:
        - bearerAuth: []
      tags: ["media"]
      summary: get photo
      description: Returns the photo file according to the path, if user authenticated is not banned by username in path
      operationId: getPhoto
      responses:
        "200":
          description: Photo returned successfully
          content:
            image/png:
              schema: {$ref: "#/components/schemas/photoFile"}
        "400": {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
        "500": {$ref: "#/components/responses/InternalServerError"}
    delete:
      security:
        - bearerAuth: []
      tags: ["media"]
      summary: delete photo
      description: |-
        If user authenticated matches username in path and photo-id matches a photo uploaded by username then deletes it.
      operationId: deletePhoto
      responses:
        "204":
          description: Photo deleted successfully
        "400": {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
        '404': {$ref: '#/components/responses/NotFound'}
        "500": {$ref: "#/components/responses/InternalServerError"} 
...
  /users/{username}/media/{photo-id}/comments/:
    parameters:
      - $ref: "#/components/parameters/photo-id"
      - $ref: "#/components/parameters/username"
    get:
      tags: ["comments"]
      security:
        - bearerAuth: []
      summary: Returns comments
      description: Returns a list of comments objects relative to the photo with photo-id given, 
        posted by username in path; if user authenticated is not banned by username.
      operationId: getComments
      responses:
        "200":
          description: List is returned successfully.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/commentArray"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: '#/components/responses/UnauthorizedError'}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalServerError"}
    
    post:
      security:
        - bearerAuth: []
      tags: ["comments"]
      summary: Comments the photo
      description: |- 
        The user authenticated posts a comment to the photo in path posted by username with 
        text given in content; if user authenticated is not banned by username. Returns the comment id.
      operationId: commentPhoto
      requestBody:
        description: The text of the comment (max 250 chars)
        content:
          application/json:
            schema: 
              type: object
              description: The text of the comment
              properties:
                text: {$ref: "#/components/schemas/commentText"}
        required: true
      responses:
        "201":
          description: Commented successfully.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/id"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: '#/components/responses/UnauthorizedError'}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalServerError"}
...
  /users/{username}/ban/{userBan}/:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/userBan"
    put:
      security:
        - bearerAuth: []
      tags: ["ban"]
      summary: ban a user
      description: |-
        User authenticated {username} bans username {userBan}, if username is not equal to userBan and
        if it doesn't exist returns an error.
      operationId: banUser
      responses:
        "201":
          description: Banned successfully
          content:
            application/json:
              schema: {$ref: "#/components/schemas/username"}        
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: '#/components/responses/UnauthorizedError'}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalServerError"}

    delete:
      tags: ["ban"]
      summary: unban
      security:
        - bearerAuth: []
      description: |-
        If {userBan} is banned by the account authenticated {username} then is unbanned.
      operationId: unbanUser
      responses:
        "204":
          description: Unbanned successfully     
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: '#/components/responses/UnauthorizedError'}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalServerError"}
...